+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to `laravel-specification-pattern` will be documented in this file.
 
+## Unreleased
+
+ ### Added
+
+- Ability to make specification verbose using composite specification.
+
 ## [v2.2.0] - 2024-03-13
 
 ### Added